pyproject.toml
[project]
name = "manipdf"
version = "0.1.0"
description = "A local, privacy-first PDF manipulation suite."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mypy>=1.20.0",
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.10.0",
    "pyside6>=6.11.0",
    "pytesseract>=0.3.13",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
    "typer>=0.24.1",
]

[project.scripts]
manipdf = "manipdf.cli.main:app"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = true
